# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RPGアイテム取引マーケットプレイスのAPI",
        "title": "RPG Market API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/items": {
            "get": {
                "description": "販売中のアイテム一覧を取得する",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "アイテム一覧取得",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "取得件数",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "オフセット",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/presenter.ItemResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "新しいアイテムを出品する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "アイテム作成",
                "parameters": [
                    {
                        "description": "アイテム情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.CreateItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/presenter.ItemResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/items/category/{category_id}": {
            "get": {
                "description": "指定したカテゴリのアイテム一覧を取得する",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "カテゴリ別アイテム一覧取得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "カテゴリID",
                        "name": "category_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "取得件数",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "オフセット",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/presenter.ItemResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/items/search": {
            "get": {
                "description": "キーワードでアイテムを検索する",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "アイテム検索",
                "parameters": [
                    {
                        "type": "string",
                        "description": "検索キーワード",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "取得件数",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "オフセット",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/presenter.ItemResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/items/seller/{seller_id}": {
            "get": {
                "description": "指定した出品者のアイテム一覧を取得する",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "出品者のアイテム一覧取得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "出品者ID",
                        "name": "seller_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "取得件数",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "オフセット",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/presenter.ItemResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/items/{id}/transaction": {
            "get": {
                "description": "自分が関わる取引情報を取得（購入者または出品者のみ）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "商品の取引情報取得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Transaction"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.ItemCondition": {
            "type": "string",
            "enum": [
                "new",
                "like_new",
                "very_good",
                "good",
                "acceptable"
            ],
            "x-enum-varnames": [
                "ItemConditionNew",
                "ItemConditionLikeNew",
                "ItemConditionVeryGood",
                "ItemConditionGood",
                "ItemConditionAcceptable"
            ]
        },
...
        "entity.ShippingDays": {
            "type": "string",
            "enum": [
                "1-2",
                "2-3",
                "4-7"
            ],
            "x-enum-varnames": [
                "ShippingDays1to2",
                "ShippingDays2to3",
                "ShippingDays4to7"
            ]
        },
...
        "presenter.ItemImageResponse": {
            "type": "object",
            "properties": {
                "display_order": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                }
            }
        },
        "presenter.ItemResponse": {
            "type": "object",
            "properties": {
                "brand_id": {
                    "type": "integer"
                },
                "category_id": {
                    "type": "integer"
                },
                "comments_count": {
                    "type": "integer"
                },
                "condition": {
                    "$ref": "#/definitions/entity.ItemCondition"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.ItemImageResponse"
                    }
                },
                "likes_count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "prefecture_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "seller": {
                    "$ref": "#/definitions/presenter.UserResponse"
                },
                "seller_id": {
                    "type": "integer"
                },
                "shipping_days": {
                    "$ref": "#/definitions/entity.ShippingDays"
                },
                "shipping_method_id": {
                    "type": "integer"
                },
                "shipping_payer": {
                    "$ref": "#/definitions/entity.ShippingPayer"
                },
                "status": {
                    "$ref": "#/definitions/entity.ItemStatus"
                },
                "stock": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "view_count": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}